Add Buttons component
