+++ public/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <title>Calculator</title>
-    <script src="index.js"></script>
+    <script src="./src/index.js"></script>
     <script src="https://kit.fontawesome.com/ea7ef19d2b.js" crossorigin="anonymous"></script>
   </head>
   <body>
